pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swegon-mcp"
version = "0.1.0"
description = "MCP server for controlling Swegon WISE ventilation systems via Modbus TCP"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pymodbus>=3.6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
swegon-mcp = "swegon_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]
